Completed sign in page
The body is divided into two subsections, a cotainer for the image, logo
and credits, and a container for the form and text.

The form container and focused form controls have a slight box shadow.
The form controls within the form also have a blue outline when focused
and invalid fields have a red outline be default.

Javascript is used to verify that the both passwords matach, if
they do not match, a psuedo element is added below the password field.